+++ blockchain/beacon/Cargo.toml
@@ -21,7 +21,6 @@ forest_shim.workspace = true
 forest_utils.workspace = true
 fvm_shared = { workspace = true, default-features = false }
 hex.workspace = true
-num_cpus.workspace = true
 parking_lot.workspace = true
 serde = { workspace = true, features = ["derive"] }
 sha2 = { workspace = true, default-features = false }
